Make settings a screen instead of a pop-over. Add the ability to pull a model. Fix a scrolling bug.
